Add book delete modal component  (#147)
 (#)
